distage-framework: almost compiles, closes #1290
