+++ distage/distage-framework/.js/src/main/scala/izumi/distage/roles/RoleAppBootConfigModule.scala
@@ -1,0 +1,12 @@
+package izumi.distage.roles
+
+import izumi.distage.config.model.AppConfig
+import izumi.distage.framework.services.ConfigLoader
+import izumi.distage.model.definition.ModuleDef
+import izumi.distage.modules.DefaultModule
+import izumi.reflect.TagK
+
+class RoleAppBootConfigModule[F[_]: TagK: DefaultModule]() extends ModuleDef {
+  make[ConfigLoader].todo
+  make[AppConfig].todo
+}
