+++ distage/distage-framework/src/main/scala/izumi/distage/roles/bundled/BundledTask.scala
