+++ distage/distage-framework/src/main/scala/izumi/distage/roles/launcher/AbstractActivationParser.scala
@@ -1,0 +1,7 @@
+package izumi.distage.roles.launcher
+
+import izumi.distage.model.definition.Activation
+
+trait AbstractActivationParser {
+  def parseActivation(): Activation
+}
